Enum msgpack_rpc::Value

pub enum Value {
    Nil,
    Boolean(bool),
    Integer(Integer),
    F32(f32),
    F64(f64),
    String(Utf8String),
    Binary(Vec<u8>),
    Array(Vec<Value>),
    Map(Vec<(Value, Value)>),
    Ext(i8, Vec<u8>),
}

Represents any valid MessagePack value.

Variants
Nil

Nil represents nil.

Boolean(bool)

Boolean represents true or false.

Integer(Integer)

Integer represents an integer.

A value of an Integer object is limited from -(2^63) upto (2^64)-1.

Examples

use rmpv::Value;

assert_eq!(42, Value::from(42).as_i64().unwrap());

F32(f32)

A 32-bit floating point number.

F64(f64)

A 64-bit floating point number.

String(Utf8String)

String extending Raw type represents a UTF-8 string.

Note

String objects may contain invalid byte sequence and the behavior of a deserializer depends on the actual implementation when it received invalid byte sequence. Deserializers should provide functionality to get the original byte array so that applications can decide how to handle the object

Binary(Vec<u8>)

Binary extending Raw type represents a byte array.

Array(Vec<Value>)

Array represents a sequence of objects.

Map(Vec<(Value, Value)>)

Map represents key-value pairs of objects.

Ext(i8, Vec<u8>)

Extended implements Extension interface: represents a tuple of type information and a byte array where type information is an integer whose meaning is defined by applications.

Methods

impl Value

fn is_nil(&self) -> bool

Returns true if the Value is a Null. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::Nil.is_nil());

fn is_bool(&self) -> bool

Returns true if the Value is a Boolean. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::Boolean(true).is_bool());

assert!(!Value::Nil.is_bool());

fn is_i64(&self) -> bool

Returns true if the Value is convertible to an i64. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::from(42).is_i64());

assert!(!Value::from(42.0).is_i64());

fn is_u64(&self) -> bool

Returns true if the Value is convertible to an u64. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::from(42).is_u64());

assert!(!Value::F32(42.0).is_u64());
assert!(!Value::F64(42.0).is_u64());

fn is_f32(&self) -> bool

Returns true if (and only if) the Value is a f32. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::F32(42.0).is_f32());

assert!(!Value::from(42).is_f32());
assert!(!Value::F64(42.0).is_f32());

fn is_f64(&self) -> bool

Returns true if (and only if) the Value is a f64. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::F64(42.0).is_f64());

assert!(!Value::from(42).is_f64());
assert!(!Value::F32(42.0).is_f64());

fn is_number(&self) -> bool

Returns true if the Value is a Number. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::from(42).is_number());
assert!(Value::F32(42.0).is_number());
assert!(Value::F64(42.0).is_number());

assert!(!Value::Nil.is_number());

fn is_str(&self) -> bool

Returns true if the Value is a String. Returns false otherwise.

Examples

use rmpv::Value;

assert!(Value::String("value".into()).is_str());

assert!(!Value::Nil.is_str());

fn is_bin(&self) -> bool

Returns true if the Value is a Binary. Returns false otherwise.

fn is_array(&self) -> bool

Returns true if the Value is an Array. Returns false otherwise.

fn is_map(&self) -> bool

Returns true if the Value is a Map. Returns false otherwise.

fn is_ext(&self) -> bool

Returns true if the Value is an Ext. Returns false otherwise.

fn as_bool(&self) -> Option<bool>

If the Value is a Boolean, returns the associated bool. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some(true), Value::Boolean(true).as_bool());

assert_eq!(None, Value::Nil.as_bool());

fn as_i64(&self) -> Option<i64>

If the Value is an integer, return or cast it to a i64. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some(42i64), Value::from(42).as_i64());

assert_eq!(None, Value::F64(42.0).as_i64());

fn as_u64(&self) -> Option<u64>

If the Value is an integer, return or cast it to a u64. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some(42u64), Value::from(42).as_u64());

assert_eq!(None, Value::from(-42).as_u64());
assert_eq!(None, Value::F64(42.0).as_u64());

fn as_f64(&self) -> Option<f64>

If the Value is a number, return or cast it to a f64. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some(42.0), Value::from(42).as_f64());
assert_eq!(Some(42.0), Value::F32(42.0f32).as_f64());
assert_eq!(Some(42.0), Value::F64(42.0f64).as_f64());

assert_eq!(Some(2147483647.0), Value::from(i32::max_value() as i64).as_f64());

assert_eq!(None, Value::Nil.as_f64());

fn as_str(&self) -> Option<&str>

If the Value is a String, returns the associated str. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some("le message"), Value::String("le message".into()).as_str());

assert_eq!(None, Value::Boolean(true).as_str());

fn as_slice(&self) -> Option<&[u8]>

If the Value is a Binary or a String, returns the associated slice. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some(&[1, 2, 3, 4, 5][..]), Value::Binary(vec![1, 2, 3, 4, 5]).as_slice());

assert_eq!(None, Value::Boolean(true).as_slice());

fn as_array(&self) -> Option<&Vec<Value>>

If the Value is an Array, returns the associated vector. Returns None otherwise.

Examples

use rmpv::Value;

let val = Value::Array(vec![Value::Nil, Value::Boolean(true)]);

assert_eq!(Some(&vec![Value::Nil, Value::Boolean(true)]), val.as_array());

assert_eq!(None, Value::Nil.as_array());

fn as_map(&self) -> Option<&Vec<(Value, Value)>>

If the Value is a Map, returns the associated vector of key-value tuples. Returns None otherwise.

Note

MessagePack represents map as a vector of key-value tuples.

Examples

use rmpv::Value;

let val = Value::Map(vec![(Value::Nil, Value::Boolean(true))]);

assert_eq!(Some(&vec![(Value::Nil, Value::Boolean(true))]), val.as_map());

assert_eq!(None, Value::Nil.as_map());

fn as_ext(&self) -> Option<(i8, &[u8])>

If the Value is an Ext, returns the associated tuple with a ty and slice. Returns None otherwise.

Examples

use rmpv::Value;

assert_eq!(Some((42, &[1, 2, 3, 4, 5][..])), Value::Ext(42, vec![1, 2, 3, 4, 5]).as_ext());

assert_eq!(None, Value::Boolean(true).as_ext());