Struct msgpack_rpc::Client

pub struct Client { /* fields omitted */ }

A client of Msgpack-RPC

Methods

impl Client

fn new(
handle: &Handle,
tx_req: UnboundedSender<(u64, Request)>,
rx_res: UnboundedReceiver<(u64, Response)>,
tx_not: UnboundedSender<(Notification, Sender<()>)>
) -> Self

Create a new Client with background task spawned on an event loop of handle.

fn request<S: Into<String>, P: Into<Value>>(
&self,
method: S,
params: P
) -> Response

Send a request message to the server, and return a future of its response.

fn notify<S: Into<String>, P: Into<Value>>(&self, method: S, params: P) -> Ack

Send a notification message to the server.